settings: Fix fallback avatars for image uploaders.
Match the behavior of the user list for loading user and organization
avatar images: if the image does not load, show a grey fallback
background color.

Before this change, the user avatar image uploader would be missing if
the user was using uBlock Origin on gravatar. The organization avatar
image uploader was present and functioning, but the background color was
a darker grey.

In this change, we attach the necessary styling, including
background color and size, to a new surrounding `div` tag instead of the
`img`, in the image uploader template. We specifically apply the style
change to the user avatar and realm icon, but not the two realm logos.
(These are the four places the template is used).

We also fix a bug wherein replacing a gravatar URL with an uploaded
image doesn't show up until the user reloads the page, due to uBlock
Origin hiding the img tag.

Fixes part of #19123.